% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_homer_output.R
\name{read_homer_output}
\alias{read_homer_output}
\title{Read known and novel motif outputs from a HOMER output folder.}
\usage{
read_homer_output(folder = "", max_files = 100)
}
\description{
Read known and novel motif outputs from a HOMER output folder.
}
