% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_homer_output.R
\name{read_homerResults_html}
\alias{read_homerResults_html}
\title{Read homerResults.html file produced by HOMER.}
\usage{
read_homerResults_html(file)
}
\description{
Read homerResults.html file produced by HOMER.
}
